sprocket 0.18.0
`sprocket` is autobumped but the workflow failed to update to version
0.18.0 because livecheck is incorrectly returning 0.19.0 as newest
(from a `wdl-v0.19.0` tag). This updates the version and adds a
`livecheck` block that restricts matching to tags like `v0.18.0`.

Upstream uses a Git submodule for `crankshaft`, so this updates the
formula to check out the Git tag. Besides that, the test output has
slightly changed, so I updated the test string accordingly.