zabbix-cli 3.5.3
`zabbix-cli` is autobumped but the workflow failed to update to
version 3.5.3 with an "Unable to update resource blocks for
'zabbix-cli' automatically. Please update them manually." error. This
manually updates the formula using the package versions from the
autobump output.

In the process, this updates the `typer` resource version, as the
related `typer.rich_utils` bug was fixed in the `zabbix-cli` 3.5.3
release.