libxmlsec1 1.3.8
`libxmlsec1` is autobumped but the workflow failed to update to
version 1.3.8 because livecheck gave an "Unable to get versions"
error in the autobump run. This manually updates the version.

The livecheck error appears to be a persistent failure in the autobump
environment, so this also updates the `livecheck` block to check
GitHub releases as a workaround.