pypy3.10: update livecheck
The existing `livecheck` block for `pypy3.10` only restricts matching
to 3.x Python versions but this formula is specifically intended for
3.10. This worked fine for a period of time but now livecheck is
returning 7.3.20 as newest and there aren't assets for Python 3.10
(only 3.11), so this isn't an appropriate version for this formula.
This updates the `livecheck` block regex to only match versions from
tarballs for Python 3.10, so it correctly returns 7.3.19 as the
latest version.