inform6: update livecheck
The existing `livecheck` block checks a directory listing page where
the `stable` archived is found but this URL now returns a 404 (Not
Found) error. It seems that upstream no longer offers this directory
listing page functionality but the website offers the same information
through their directory index pages. This updates the `livecheck`
block URL to check the index for the directory, which links to the
`stable` tarball in the same way as before.