perl: update livecheck
The existing `livecheck` block for `perl` checks the upstream Perl
Source page and this previously linked to the newest tarballs but now
it doesn't link to any tarballs at all. This addresses the issue by
updating the `livecheck` block URL to check the directory listing page
for the version directory where the `stable` tarball is found. This
approach will fail to surface new major versions (i.e., it will
continue to return the newest 5.x version), so we will have to
manually identify new major versions (which should be very rare).