libdecor 0.2.4 (new formula)
Add libdecor, a library for wayland on linux that helps applications
running on GNOME draw window decorations.

This is beneficial on Fedora Silverblue and derivatives for providing
pkg-config files for libdecor. Fedora Silverblue already ships with
the library, but it doesn't have the development packages by default,
and since it is an immutable OS, it is preferential to acquire them
through brew, much like the wayland formulae.