asyncapi: bump version_scheme
We recently downgraded `asyncapi` from 4.1.3 to 4.1.1, as 4.1.3 was
yanked (presumably due to Shai Hulud 2.0). This ensured that new
installations and or upgrades afterward would install 4.1.1 but
existing users who had upgraded to 4.1.3 would not be automatically
downgraded. This bumps the `version_scheme` for the formula to ensure
that users on 4.1.3 will be automatically downgraded.