rv: remove now-unneeded openssl dependency
rv no longer depends on openssl, as a side effect of a pull request
adding support for arch linux. that means we can remove this dependency.

also remove install command. now that the run command automatically
installs missing ruby versions, we don't need to explicitly install
anything for the last command to succeed.